chat-server/requirements.txt
aiofiles==23.2.1
bidict==0.22.1
blinker==1.6.3
click==8.1.7
dnspython==2.4.2
eventlet==0.33.3
Flask==3.0.0
gevent==23.9.1
greenlet==3.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.14.4
hyperframe==6.0.1
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mqtt==0.0.1
paho-mqtt==1.6.1
priority==2.0.0
python-engineio==4.8.0
python-socketio==5.10.0
Quart==0.19.3
simple-websocket==1.0.0
six==1.16.0
tomli==2.0.1
typing_extensions==4.8.0
websocket-client==1.6.4
Werkzeug==3.0.0
wsproto==1.2.0
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
